moving AdvocateSearch into a features sub-folder under components
