+++ src/app/page.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useState, useMemo } from 'react';
 
-import { AdvocatesSearch } from '@/components/AdvocatesSearch';
+import { AdvocatesSearch } from '@/components/features/AdvocatesSearch';
 import { AdvocatesTable } from '@/components/AdvocatesTable';
 import { PgTitle } from '@/components/PgTitle';
